refactor(parser): extract module item dispatch
